// generated/src/app/api-models/claim-audit-result-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimAuditResultDetail {
  claimResultCode: string;
  editType: string;
  editValue: string;
  editOrder: number;
  codeOrigination: number;
  claimAction: string;
  reasnCode: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}